Reduce chance of crashes in CTRL-C handler, rework handler example
Inside a signal handler, you cannot allocate memory because the signal
handler, being implemented with a C
[`signal`](https://en.cppreference.com/w/c/program/signal) call, can be
called _during_ a memory allocation - when that happens, the CTRL-C
handler causes a segfault and/or other inconsistent state.

Similarly, the call can happen from a non-nim thread or inside a C
library function call etc, most of which do not support reentrancy and
therefore cannot be called _from_ a signal handler.

The stack trace facility used in the default handler is unfortunately
beyond fixing without more significant refactoring since it uses
garbage-collected types in its API and implementation, but we can at
least allocate the buffer outside of the signal handler itself -
hopefully, this should reduce the frequency of crashes, if nothing else.

It will still crash from time to time on Windows in particular, but
since we're about to quit without cleanup, the loss of functionality is
simited (ie the stack trace will not show and a crash dump will happen
which the OS will notice).

Finally, the example of a ctrl-c handler performs the same mistake of
calling `echo` which is not well-defined - replace it with an example
that is mostly correct (except maybe for the lack of `volatile` for the
`stop` variable).

eh

skip signal handler in alloc-counting tests

more alloc stats

raises?

fun with raises

more tests

reenable on windows

we're about to terminate anyway