Do not directly cast int128 to uint64 in semfold
int128 is an array of uint32s, so while this works on little-endian CPUs,
it's completely broken on big-endian.  e.g. following snippet would fail:

	const x = 0xFFFFFFFF'u32
	const y = (x shr 1)
	echo y # amd64: 2147483647, s390x: 0

That in turn broke float printing, resulting in miscompilation of any code
that used floats.

To fix this, we now call the aptly named castToUInt64 procedure which
performs the same cast portably.

(Thanks to barracuda156 for helping debug this.)